Launcher_Android: ignore ALOOPER_POLL_WAKE when application is in background
When the application is in the background, an ALOOPER_POLL_WAKE message may arrive, usually when opening another application.
This would cause an exit from the wait in `looperFunc` and continue the execution of `MainLoop` and `gameApplication.Tick()`.

This change modifies the loop to ignore `ALOOPER_POLL_WAKE` messages when `androidEnv->IsRunning()` is false, ensuring the application remains paused in the background.